feat(user-api): :sparkles: implement user deletion endpoint with error handling and logging
