+++ Apps/Accounts/api/__init__.py
@@ -1,0 +1,1 @@
+from .user import *
