+++ Apps/Accounts/api/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path, include
+
+urlpatterns = [
+    path('user/', include('Apps.Accounts.api.user.urls')),
+
+]
