+++ Apps/Accounts/urls.py
@@ -1,6 +1,7 @@
-from django.urls import path
-from Apps.Accounts.views import RegisterUserView, UpdateUserView, UserDeleteView, GetUserView
+from django.urls import path, include
 from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
+
+from Apps.Accounts.views import RegisterUserView, UpdateUserView, GetUserView
 
 app_name = 'accounts'
 urlpatterns = []
@@ -9,11 +10,11 @@ urlpatterns += [
     # User
     path('register/', RegisterUserView.as_view(), name='register'),
     path('update/', UpdateUserView.as_view(), name='user_update'),
-    path('delete/', UserDeleteView.as_view(), name='delete'),
 
     # Api
     path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
     path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
     path('api/me/', GetUserView.as_view(), name='get_user'),
 
+    path('api/', include('Apps.Accounts.api.urls'))
 ]
